requirements.txt
streamlit
transformers
google-cloud-bigquery
